moves engine definition into the main require.
- this is part of the simplification process, and will eventually get routes moved into it.
