+++ lib/jasminerice.rb
@@ -1,5 +1,3 @@
-require "jasminerice/engine"
-
 module Jasminerice
   # Determine whether or not to mount the Jasminerice engine implicitly. True/False
   mattr_accessor :mount
@@ -17,5 +15,20 @@ module Jasminerice
   # a fresh initializer with all configuration values.
   def self.setup
     yield self
+  end
+
+  class Engine < Rails::Engine
+    isolate_namespace Jasminerice
+
+    initializer :assets, :group => :all do |app|
+      app.config.assets.paths << Rails.root.join("spec", "javascripts")
+      app.config.assets.paths << Rails.root.join("spec", "stylesheets")
+    end
+
+    config.after_initialize do |app|
+      app.routes.prepend do
+        mount Jasminerice::Engine => Jasminerice.mount_at
+      end if Jasminerice.mount
+    end
   end
 end
